Moved Option & SlashCommand annotations to annotations.types
