+++ src/main/java/joseta/annotations/CommandProcessor.java
@@ -1,6 +1,7 @@
 package joseta.annotations;
 
 import joseta.annotations.modules.*;
+import joseta.annotations.types.*;
 import joseta.utils.*;
 import net.dv8tion.jda.api.*;
 import net.dv8tion.jda.api.entities.*;
@@ -10,7 +11,6 @@ import net.dv8tion.jda.api.events.interaction.command.*;
 import net.dv8tion.jda.api.hooks.*;
 import net.dv8tion.jda.api.interactions.commands.*;
 import net.dv8tion.jda.api.interactions.commands.build.*;
-import org.jetbrains.annotations.*;
 import org.reflections.*;
 
 import java.lang.reflect.*;
