+++ src/main/java/joseta/annotations/types/Option.java
@@ -1,4 +1,4 @@
-package joseta.annotations;
+package joseta.annotations.types;
 
 import java.lang.annotation.*;
