+++ src/main/java/joseta/commands/MiscCommands.java
@@ -1,7 +1,7 @@
 package joseta.commands;
 
-import joseta.annotations.*;
 import joseta.annotations.modules.*;
+import joseta.annotations.types.*;
 import net.dv8tion.jda.api.entities.*;
 import net.dv8tion.jda.api.events.interaction.command.*;
